cnwave_network_analyst/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
branca==0.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
DateTime==5.5
distro==1.9.0
dotenv==0.9.9
folium==0.20.0
functions==0.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
logging==0.4.9.6
MarkupSafe==3.0.2
numpy==2.3.1
openai==1.93.0
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.4.0
zope.interface==7.2
